refactor: moved `StoreProvider` inside contexts
